refactor(Data Cnnnections): move /pages to /tabs
