+++ public/app/features/data-connections/tabs/CloudIntegrations/CloudIntegrations.tsx
